// src/libs/types/callings.ts
export interface Organization {
  unitNumber: number
  unitType: number
  subOrgId: number
  parentSubOrgId: any
  parentName?: string
  customOrgName?: string
  defaultOrgName: string
  name: string
  relatedSubOrgName: any
  priesthood: boolean
  gender?: string
  adultClass: boolean
  filterOffices: FilterOffice[]
  canHaveMembers: boolean
  userCanEditCallings: boolean
  userCanEditClassAssignments: boolean
  userCanSeeBasicInfo: boolean
  multipleAllowed: boolean
  nameEditable: boolean
  resetable: boolean
  required: boolean
  customPositionsAllowed: boolean
  missionaries: boolean
  highCouncilOrDistrictCouncil: boolean
  room: any
  splitOrCombinedClassNotice?: string
  canShowWithNoCallings: boolean
  children: Children[]
  displayOrder: number
  instance: number
  totalInstances: number
  callings: Calling[]
  members: any[]
  size: any
  firstOrgTypeId: number
  firstOrgTypeName: any
  orgTypeIds: number | undefined[]
  addableSubOrgTypes: any
  positionTypeIdsThatCanHaveResponsibilities: any[]
  isSplit: boolean
  leadershipSubOrgId: any
  classes: any
  mergeableSubOrgTypeIds?: any[]
  classesCollapsed: boolean
  other: boolean
  enableHT: boolean
  enableVT: boolean
  supportsAttendanceRoll: boolean
  hasCustomCallings: boolean
  hasAnyAllowMultiple: boolean
  firstOrgType: string
  classGroup?: string
}

export interface FilterOffice {
  codes: number | undefined[]
  name: string
}

export interface Children {
  unitNumber: number
  unitType: number
  subOrgId: number
  parentSubOrgId: number
  parentName: string
  customOrgName?: string
  defaultOrgName: string
  name: string
  relatedSubOrgName: any
  priesthood: boolean
  gender?: string
  adultClass: boolean
  filterOffices: FilterOffice[]
  canHaveMembers: boolean
  userCanEditCallings: boolean
  userCanEditClassAssignments: boolean
  userCanSeeBasicInfo: boolean
  multipleAllowed: boolean
  nameEditable: boolean
  resetable: boolean
  required: boolean
  customPositionsAllowed: boolean
  missionaries: boolean
  highCouncilOrDistrictCouncil: boolean
  room?: string
  splitOrCombinedClassNotice?: string
  canShowWithNoCallings: boolean
  children: Children[]
  displayOrder: number
  instance: number
  totalInstances: number
  callings: Calling[]
  members: any[]
  size: any
  firstOrgTypeId: number
  firstOrgTypeName: any
  orgTypeIds: number[]
  addableSubOrgTypes: any
  positionTypeIdsThatCanHaveResponsibilities: any[]
  isSplit: boolean
  leadershipSubOrgId: any
  classes: any
  mergeableSubOrgTypeIds: number[]
  classesCollapsed: boolean
  other: boolean
  enableHT: boolean
  enableVT: boolean
  supportsAttendanceRoll: boolean
  hasCustomCallings: boolean
  hasAnyAllowMultiple: boolean
  firstOrgType: string
  classGroup?: string
}

export interface OtherAllowedSubOrg {
  unitNumber: number
  subOrgId: number
  orgTypeIds: number[]
  defaultOrgTypeIds: number[]
  name: string
  children: any[]
  userCanEditCallings: boolean
  isClass: boolean
  isRealClass: boolean
  isSplit: boolean
  firstTypeId: number
  gender?: string
  isCombined: boolean
  parentName?: string
  classGroup?: string
}

export interface ContactMailingAddr {
  street1: string
  street2: any
  city: string
  district: any
  postalCode: string
  state: number
  country: number
  formattedLines: any[]
  standardized: any
  geocodeToUnits: boolean
  units: any
}

export interface Calling {
  unitNumber?: number
  unitName: any
  subOrgTypeId: number
  subOrgId: number
  newSubOrgId: any
  organization: string
  parentOrganization: any
  otherAllowedSubOrgs?: OtherAllowedSubOrg[]
  positionTypeId?: number
  positionId?: number
  releasePositionIds: any
  position: string
  positionDisplayOrder?: number
  memberId?: number
  mrn?: string
  nameVerificationForMrn: any
  memberName?: string
  memberNameOrder: number
  memberEmail?: string
  memberPhone?: string
  memberUnitName?: string
  memberUnitNumber?: number
  setApartByPersonId?: number
  setApartByPersonMrn?: string
  setApartByPersonName?: string
  ordainedByPersonId?: number
  ordainedByPersonMrn?: string
  ordainedByPersonName?: string
  activeDate?: string
  expectedEndDate?: string
  releaseDate: any
  setApartDate?: string
  ordainDate?: string
  missionName?: string
  missionUnitNumber?: number
  missionaryHomeUnitName: any
  dualAuthUser: any
  dualAuthPass: any
  responsibilities: any
  contactWorkPhone: any
  contactMobilePhone?: string
  contactEmail?: string
  contactMailingAddr?: ContactMailingAddr
  autoReleasedPositionTypeIds: any
  editable: boolean
  contactInfoEditable: boolean
  allowMultiple: boolean
  custom: boolean
  setApart: boolean
  customNameChanged: boolean
  hidden: boolean
  allowNew: boolean
  userCanEdit: boolean
  userCanEditContactInfoOnly: boolean
  canHaveResponsibilities: boolean
  unpublished: boolean
  requiresSetApartDateAndPerson: boolean
  requiresBishopOrdination: boolean
  allowVacant: boolean
  requiresDualAuthorization: boolean
  pendingDelete: boolean
  hasContactInfo: boolean
  showUnitCard: boolean
  pendingCalling: boolean
  pendingDualAuth: boolean
  allowApprove: boolean
  previouslyOrdained: boolean
  userCannotEditPreviousOrdination: boolean
  approvalRequest: boolean
  sequence: number
  missionary: boolean
  vacant: boolean
  positionTypeEnum?: string
}